Distribute test_step into step_methods
