+++ pymc/tests/step_methods/hmc/test_nuts.py
@@ -12,9 +12,20 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
+import sys
+import warnings
+
+import aesara.tensor as at
+import numpy as np
 import pytest
 
+import pymc as pm
+
+from pymc.aesaraf import floatX
+from pymc.exceptions import SamplingError
+from pymc.step_methods.hmc import NUTS
 from pymc.tests import sampler_fixtures as sf
+from pymc.tests.helpers import RVsAssignmentStepsTester, StepMethodTester
 
 
 class TestNUTSUniform(sf.NutsFixture, sf.UniformFixture):
@@ -81,3 +92,116 @@ class TestNUTSLKJCholeskyCov(sf.NutsFixture, sf.LKJCholeskyCovFixture):
     burn = 0
     chains = 2
     min_n_eff = 200
+
+
+class TestNutsCheckTrace:
+    def test_multiple_samplers(self, caplog):
+        with pm.Model():
+            prob = pm.Beta("prob", alpha=5.0, beta=3.0)
+            pm.Binomial("outcome", n=1, p=prob)
+            caplog.clear()
+            with warnings.catch_warnings():
+                warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+                pm.sample(3, tune=2, discard_tuned_samples=False, n_init=None, chains=1)
+            messages = [msg.msg for msg in caplog.records]
+            assert all("boolean index did not" not in msg for msg in messages)
+
+    def test_bad_init_nonparallel(self):
+        with pm.Model():
+            pm.HalfNormal("a", sigma=1, initval=-1, transform=None)
+            with pytest.raises(SamplingError) as error:
+                pm.sample(chains=1, random_seed=1)
+            error.match("Initial evaluation")
+
+    @pytest.mark.skipif(sys.version_info < (3, 6), reason="requires python3.6 or higher")
+    def test_bad_init_parallel(self):
+        with pm.Model():
+            pm.HalfNormal("a", sigma=1, initval=-1, transform=None)
+            with pytest.raises(SamplingError) as error:
+                pm.sample(cores=2, random_seed=1)
+            error.match("Initial evaluation")
+
+    def test_linalg(self, caplog):
+        with pm.Model():
+            a = pm.Normal("a", size=2, initval=floatX(np.zeros(2)))
+            a = at.switch(a > 0, np.inf, a)
+            b = at.slinalg.solve(floatX(np.eye(2)), a, check_finite=False)
+            pm.Normal("c", mu=b, size=2, initval=floatX(np.r_[0.0, 0.0]))
+            caplog.clear()
+            with warnings.catch_warnings():
+                warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+                trace = pm.sample(20, tune=5, chains=2, return_inferencedata=False, random_seed=526)
+            warns = [msg.msg for msg in caplog.records]
+            assert np.any(trace["diverging"])
+            assert (
+                any("divergence after tuning" in warn for warn in warns)
+                or any("divergences after tuning" in warn for warn in warns)
+                or any("only diverging samples" in warn for warn in warns)
+            )
+
+            with pytest.raises(ValueError) as error:
+                trace.report.raise_ok()
+            error.match("issues during sampling")
+
+            assert not trace.report.ok
+
+    def test_sampler_stats(self):
+        with pm.Model() as model:
+            pm.Normal("x", mu=0, sigma=1)
+            with warnings.catch_warnings():
+                warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+                trace = pm.sample(draws=10, tune=1, chains=1, return_inferencedata=False)
+
+        # Assert stats exist and have the correct shape.
+        expected_stat_names = {
+            "depth",
+            "diverging",
+            "energy",
+            "energy_error",
+            "model_logp",
+            "max_energy_error",
+            "mean_tree_accept",
+            "step_size",
+            "step_size_bar",
+            "tree_size",
+            "tune",
+            "perf_counter_diff",
+            "perf_counter_start",
+            "process_time_diff",
+            "index_in_trajectory",
+            "largest_eigval",
+            "smallest_eigval",
+        }
+        assert trace.stat_names == expected_stat_names
+        for varname in trace.stat_names:
+            assert trace.get_sampler_stats(varname).shape == (10,)
+
+        # Assert model logp is computed correctly: computing post-sampling
+        # and tracking while sampling should give same results.
+        model_logp_fn = model.compile_logp()
+        model_logp_ = np.array(
+            [
+                model_logp_fn(trace.point(i, chain=c))
+                for c in trace.chains
+                for i in range(len(trace))
+            ]
+        )
+        assert (trace.model_logp == model_logp_).all()
+
+
+class TestStepNUTS(StepMethodTester):
+    @pytest.mark.parametrize(
+        "step_fn, draws",
+        [
+            (lambda C, _: NUTS(scaling=C, is_cov=True, blocked=False), 1000),
+            (lambda C, _: NUTS(scaling=C, is_cov=True), 1000),
+        ],
+    )
+    def test_step_continuous(self, step_fn, draws):
+        self.step_continuous(step_fn, draws)
+
+
+class TestRVsAssignmentNUTS(RVsAssignmentStepsTester):
+    @pytest.mark.parametrize("step, step_kwargs", [(NUTS, {})])
+    def test_continuous_steps(self, step, step_kwargs):
+        self.continuous_steps(step, step_kwargs)
